backend/requirements.txt

flask
flask-cors
deep-translator
werkzeug
python-dotenv
